Handle Optionals

+++ 22.SQL-databases-and-JPA/src/main/java/com/ltp/gradesubmission/service/GradeServiceImpl.java
@@ -1,16 +1,19 @@
 package com.ltp.gradesubmission.service;
 
-import java.util.List;
-import java.util.Optional;
-
 import com.ltp.gradesubmission.entity.Course;
 import com.ltp.gradesubmission.entity.Grade;
 import com.ltp.gradesubmission.entity.Student;
+import com.ltp.gradesubmission.exception.CourseNotFoundException;
+import com.ltp.gradesubmission.exception.GradeNotFoundException;
+import com.ltp.gradesubmission.exception.StudentNotFoundException;
 import com.ltp.gradesubmission.repository.CourseRepository;
 import com.ltp.gradesubmission.repository.GradeRepository;
 import com.ltp.gradesubmission.repository.StudentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class GradeServiceImpl implements GradeService {
@@ -27,17 +30,34 @@ public class GradeServiceImpl implements GradeService {
     @Override
     public Grade getGrade(Long studentId, Long courseId) {
         Optional<Grade> optionalGrade = gradeRepository.findFirstByCourseIdAndStudentId(courseId, studentId);
+
+        //no grade was found
+        if (optionalGrade.isEmpty()) {
+            throw new GradeNotFoundException(studentId, courseId);
+        }
+
+        //grade was found
         return optionalGrade.get();
     }
 
     @Override
     public Grade saveGrade(Grade grade, Long studentId, Long courseId) {
         //save a grade with the student id and course id it belongs to
-        Student student = studentRepository.findById(studentId).get();
-        Course course = courseRepository.findById(courseId).get();
 
-        grade.setStudent(student);
-        grade.setCourse(course);
+        //check if the student exists
+        Optional<Student> optionalStudent = studentRepository.findById(studentId);
+        if (optionalStudent.isEmpty()) {
+            throw new StudentNotFoundException(studentId);
+        }
+
+        //check if the course exists
+        Optional<Course> optionalCourse = courseRepository.findById(courseId);
+        if (optionalCourse.isEmpty()) {
+            throw new CourseNotFoundException(courseId);
+        }
+
+        grade.setStudent(optionalStudent.get());
+        grade.setCourse(optionalCourse.get());
 
         Grade savedGrade = gradeRepository.save(grade);
 
@@ -47,6 +67,12 @@ public class GradeServiceImpl implements GradeService {
     @Override
     public Grade updateGrade(String score, Long studentId, Long courseId) {
         Optional<Grade> optionalGrade = gradeRepository.findFirstByCourseIdAndStudentId(courseId, studentId);
+
+        //the grade to be updated does not exist
+        if (optionalGrade.isEmpty()) {
+            throw new GradeNotFoundException(studentId, courseId);
+        }
+
 
         Grade grade = optionalGrade.get();
         grade.setScore(score);
@@ -58,6 +84,13 @@ public class GradeServiceImpl implements GradeService {
     @Override
     public void deleteGrade(Long studentId, Long courseId) {
         Optional<Grade> optionalGrade = gradeRepository.findFirstByCourseIdAndStudentId(courseId, studentId);
+
+        //the grade to be deleted does not exist
+        if (optionalGrade.isEmpty()) {
+            throw new GradeNotFoundException(studentId, courseId);
+        }
+
+        //the grade exists
         Grade grade = optionalGrade.get();
         gradeRepository.deleteById(grade.getId());
     }
@@ -65,14 +98,30 @@ public class GradeServiceImpl implements GradeService {
     @Override
     public List<Grade> getStudentGrades(Long studentId) {
         Optional<List<Grade>> optionalGrades = gradeRepository.findAllByStudentId(studentId);
+
         List<Grade> grades = optionalGrades.get();
+
+        //not grade found with that sudentId
+        if (optionalGrades.isEmpty() || grades.size() == 0) {
+            throw new GradeNotFoundException(studentId, null);
+        }
+
+        //at least one grade was found
         return grades;
     }
 
     @Override
     public List<Grade> getCourseGrades(Long courseId) {
         Optional<List<Grade>> optionalGrades = gradeRepository.findAllByCourseId(courseId);
+
         List<Grade> grades = optionalGrades.get();
+
+        //no grade was found
+        if (optionalGrades.isEmpty() || grades.size() == 0) {
+            throw new GradeNotFoundException(null, courseId);
+        }
+
+        //at least one grade was found
         return grades;
     }
